existing user checked and throw error

+++ index.js
@@ -43,11 +43,22 @@ app
 
 app.post("/api/users", (req, res) => {
   const body = req.body;
+  const existingUser = users.find(
+    (user) =>
+      (user.first_name === body.first_name &&
+        user.last_name === body.last_name) ||
+      user.email === body.email
+  );
 
-  users.push({ id: users.length + 1, ...body });
-  fs.writeFile("./MOCK_DATA.json", JSON.stringify(users), (err, data) => {
-    return res.json({ status: "success", id: users.length });
-  });
+  if (existingUser) {
+    console.log(existingUser);
+    return res.json({ status: "error" }, { message: "user already exist" });
+  } else {
+    users.push({ id: users.length + 1, ...body });
+    fs.writeFile("./MOCK_DATA.json", JSON.stringify(users), (err, data) => {
+      return res.json({ status: "success", id: users.length });
+    });
+  }
 });
 app.listen(PORT, () => {
   console.log(`Server starte at PORT ${PORT}`);
